+++ js/tests/test_widget.js
@@ -1,9 +1,13 @@
-import { Select2Widget } from "../src/widget";
+import {Select2Widget} from "../src/widget";
+import {register_array_subscribers} from "../src/widget";
 
 QUnit.module('select2', hooks => {
     let wid;
     let el = $('<input class="select2" value=""/>');
     let vocab = ['One', 'Two', 'Four'];
+    let _array_subscribers = {
+        on_add: []
+    };
 
     hooks.beforeEach(()=> {
         el.appendTo('body');
@@ -28,6 +32,57 @@ QUnit.module('select2', hooks => {
         assert.ok(wid);
 
         assert.verifySteps(['select2 called']);
+    });
+
+    QUnit.test('register_array_subscribers', assert => {
+        $.fn.extend({
+            select2: function () {
+                assert.step('select2 called');
+                return this;
+            }
+        });
+
+        // return if window.yafowil === undefined
+        register_array_subscribers();
+        assert.deepEqual(_array_subscribers['on_add'], []);
+
+        // patch yafowil_array
+        window.yafowil_array = {
+            on_array_event: function(evt_name, evt_function) {
+                _array_subscribers[evt_name] = evt_function;
+            },
+            inside_template(elem) {
+                return elem.parents('.arraytemplate').length > 0;
+            }
+        };
+        register_array_subscribers();
+
+        // create table DOM
+        let table = $('<table />')
+            .append($('<tr />'))
+            .append($('<td />'))
+            .appendTo('body');
+
+        $('td', table).addClass('arraytemplate');
+        el.detach().appendTo($('td', table));
+
+        // invoke array on_add - returns
+        _array_subscribers['on_add'].apply(null, $('tr', table));
+        wid = el.data('yafowil-select2');
+        assert.notOk(wid);
+        $('td', table).removeClass('arraytemplate');
+        assert.verifySteps([]);
+
+        // invoke array on_add
+        $('td', table).removeClass('arraytemplate');
+        _array_subscribers['on_add'].apply(null, $('tr', table));
+        wid = el.data('yafowil-select2');
+        assert.ok(wid);
+        assert.verifySteps(['select2 called']);
+
+        table.remove();
+        window.yafowil_array = undefined;
+        _array_subscribers = undefined;
     });
 
     QUnit.test('fail initialize', assert => {