+++ src/yafowil/widget/select2/resources/widget.js
@@ -5,6 +5,10 @@ var yafowil_select2 = (function (exports, $) {
         static initialize(context) {
             $('.select2', context).each(function (event) {
                 let elem = $(this);
+                if (window.yafowil_array !== undefined &&
+                    window.yafowil_array.inside_template(elem)) {
+                    return;
+                }
                 let options = elem.data();
                 new Select2Widget(elem, options);
             });
@@ -78,6 +82,15 @@ var yafowil_select2 = (function (exports, $) {
             return value;
         }
     }
+    function select2_on_array_add(inst, context) {
+        Select2Widget.initialize(context);
+    }
+    function register_array_subscribers() {
+        if (window.yafowil_array === undefined) {
+            return;
+        }
+        window.yafowil_array.on_array_event('on_add', select2_on_array_add);
+    }
 
     $(function() {
         if (window.ts !== undefined) {
@@ -87,9 +100,11 @@ var yafowil_select2 = (function (exports, $) {
         } else {
             Select2Widget.initialize();
         }
+        register_array_subscribers();
     });
 
     exports.Select2Widget = Select2Widget;
+    exports.register_array_subscribers = register_array_subscribers;
 
     Object.defineProperty(exports, '__esModule', { value: true });